Workaround for broken vSphere SDK on Ubuntu >=14.04
Background Story:
Net::SSL does not support SSL validation, but IO::Socket::SSL does. On recent Ubuntu
our code stopped to work with the SDK. A deep dive showed that it was related to using
Net::SSL (which maybe has other issues on recent Ubuntu) instead of IO::Socket::SSL
and also related to SSL validation happening.

Our fix solves this problem in two steps if IO::Socket::SSL is available:
1. Force the SDK to use it.
2. Correctly disable SSL validation according to LML configuration.

As a result SSL validation should work if desired (cannot test it :-() and supported and
LML should work on older and newer systems and with older and recent SDK versions.